# Conflicts:
#	marketing-inner-api/src/main/java/com/br/marketing/innerapi/config/WebConfig.java
#	marketing-utils/src/main/java/com/br/marketing/common/constants/rediskey/RedisKeyConstant.java
#	marketing-utils/src/main/java/com/br/marketing/common/enums/TableCodeEnum.java

+++ marketing-cs/src/main/java/com/br/marketing/mapper/MarketingCustomerConfigMapperBase.java
@@ -2,11 +2,12 @@ package com.br.marketing.mapper;
 
 import com.br.marketing.entity.MarketingCustomerConfig;
 import com.br.marketing.entity.MarketingCustomerConfigExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
+import java.util.List;
+
 public interface MarketingCustomerConfigMapperBase {
-    long countByExample(MarketingCustomerConfigExample example);
+    int countByExample(MarketingCustomerConfigExample example);
 
     int deleteByExample(MarketingCustomerConfigExample example);
